Fixed double-hashing of AADClientSecret variable
Previous deployment script had AADClientSecret stored as a string and converted downstream into a securestring captured in a different variable. 

The script was double-hashing the AADClientSecret variable, leading to key vault access issues due to mismatched secret(s). 

This fix removes the unnecessary double-hashing of AADClientSecret and correctly passes in the SecureString into the keyvault. Also captures the correct client secret for use in automated ARM deployments.