Fixed double-hashing of AADClientSecret
Original pre-deployment script had the AADClientSecret stored as a string, as previous versions of the PowerShell ARM module did not require it to be a securestring parameter. 

Updated the pre-deployment script to generate an AADClientSecret as a securestring and cleaned up older code/logic that resulted in the double-hashing of that parameter.