Nesting AD VM Set Encryption
Further gating the timing of template calls. 

SQL and Management VMs were set to previously kick off after the primary DC had finished encrypting. 

As the order in which a template for any VM set is kicked off can vary with each template, explicitly setting a requirement for SQL VM encryption to not occur until all AD VMs are encrypted.