Pre-Deployment UX Updates
Added in appropriate messaging for addressing errors that could occur in pre-deployment. 
Previously, any pre-deployment errors would end in premature script termination, with no error message output. 

The latest updates now include appropriate error messages as well as recommended guidance for users on next steps to follow, if an error occurs in pre-deployment. 

Domain validation is also in the script. NetBIOS and RFC 1123 compliant domain names are the requirement for the script. 

UX has been updated to use the following:
Yellow font for progress messages
Green font for success messages 
Magenta font for in-line errors that can be corrected without ending the script
Red font for PowerShell default errors